# docs/contracts/openapi.snapshot.json
{
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "7/24 Finansal Sinyal ve Analiz REST API",
    "title": "Otonom Analiz API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/realtime/sse/signals": {
      "get": {
        "description": "SSE endpoint for trading signals.",
        "operationId": "sse_signals_realtime_sse_signals_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sse Signals",
        "tags": [
          "Real-time Data"
        ],
        "x-access-role": "admin"
      }
    },
    "/realtime/sse/ticker": {
      "get": {
        "description": "Server-Sent Events endpoint for ticker data.\nAlternative to WebSocket for simpler clients.",
        "operationId": "sse_ticker_realtime_sse_ticker_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sse Ticker",
        "tags": [
          "Real-time Data"
        ],
        "x-access-role": "admin"
      }
    },
    "/realtime/status": {
      "get": {
        "description": "Get real-time service status.",
        "operationId": "realtime_status_realtime_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Realtime Status",
        "tags": [
          "Real-time Data"
        ],
        "x-access-role": "admin"
      }
    },
...
  }
}